✅ Separate graph tests
