+++ static/app/views/issueDetails/streamline/eventDetailsHeader.spec.tsx
@@ -26,10 +26,10 @@ describe('EventDetailsHeader', () => {
   });
   const group = GroupFixture();
   const event = EventFixture({id: 'event-id'});
-  const persistantQuery = `issue:${group.shortId}`;
   const defaultProps = {group, event};
-
-  let mockEventStats: jest.Mock;
+  const router = RouterFixture({
+    location: LocationFixture({query: {streamline: '1'}}),
+  });
 
   beforeEach(() => {
     MockApiClient.clearMockResponses();
@@ -48,159 +48,37 @@ describe('EventDetailsHeader', () => {
       new Set(['environments'])
     );
     ProjectsStore.loadInitialData([project]);
-    mockEventStats = MockApiClient.addMockResponse({
+    MockApiClient.addMockResponse({
       url: `/organizations/${organization.slug}/events-stats/`,
       body: {'count()': EventsStatsFixture(), 'count_unique(user)': EventsStatsFixture()},
       method: 'GET',
     });
   });
 
-  it('displays allows toggling data sets', async function () {
-    render(<EventDetailsHeader {...defaultProps} />, {organization});
-    const eventsToggle = await screen.findByRole('button', {
-      name: 'Toggle graph series - Events',
-    });
-    const usersToggle = screen.getByRole('button', {name: 'Toggle graph series - Users'});
-
-    // Defaults to events graph
-    expect(eventsToggle).toBeDisabled();
-    expect(usersToggle).toBeEnabled();
-
-    // Switch to users graph
-    await userEvent.click(usersToggle);
-    expect(eventsToggle).toBeEnabled();
-    expect(usersToggle).toBeDisabled();
-
-    // Another click should do nothing
-    await userEvent.click(usersToggle);
-    expect(eventsToggle).toBeEnabled();
-    expect(usersToggle).toBeDisabled();
-
-    // Switch back to events
-    await userEvent.click(eventsToggle);
-    expect(eventsToggle).toBeDisabled();
-    expect(usersToggle).toBeEnabled();
-  });
-
-  it('renders the graph using a discover event stats query', async function () {
-    render(<EventDetailsHeader {...defaultProps} />, {organization});
-    await screen.findByRole('button', {
-      name: 'Toggle graph series - Events',
-    });
-
-    expect(mockEventStats).toHaveBeenCalledWith(
-      '/organizations/org-slug/events-stats/',
-      expect.objectContaining({
-        query: {
-          dataset: 'errors',
-          environment: [],
-          field: expect.anything(),
-          partial: 1,
-          interval: '12h',
-          per_page: 50,
-          project: [project.id],
-          query: persistantQuery,
-          referrer: 'issue_details.streamline_graph',
-          statsPeriod: '14d',
-          yAxis: ['count()', 'count_unique(user)'],
-        },
-      })
-    );
-
-    const discoverButton = screen.getByLabelText('Open in Discover');
-    expect(discoverButton).toBeInTheDocument();
-    expect(discoverButton).toHaveAttribute(
-      'href',
-      expect.stringContaining(`/organizations/${organization.slug}/discover/results/`)
-    );
-  });
-
-  it('allows filtering by environment', async function () {
-    render(<EventDetailsHeader {...defaultProps} />, {organization});
-    await screen.findByRole('button', {
-      name: 'Toggle graph series - Events',
-    });
-
-    await userEvent.click(screen.getByRole('button', {name: 'All Envs'}));
-    await userEvent.click(screen.getByRole('row', {name: 'production'}));
-
-    expect(mockEventStats).toHaveBeenCalledWith(
-      '/organizations/org-slug/events-stats/',
-      expect.objectContaining({
-        query: expect.objectContaining({
-          environment: ['production'],
-        }),
-      })
-    );
-  });
-
-  it('updates query from location param change', async function () {
-    const [tagKey, tagValue] = ['user.email', 'leander.rodrigues@sentry.io'];
-    const locationQuery = {
-      query: {
-        query: `${tagKey}:${tagValue}`,
-      },
-    };
-    const router = RouterFixture({
-      location: LocationFixture(locationQuery),
-    });
-
+  it('renders filters alongside the graph', async function () {
     render(<EventDetailsHeader {...defaultProps} />, {organization, router});
-    await screen.findByRole('button', {
-      name: 'Toggle graph series - Events',
-    });
+    expect(await screen.findByTestId('event-graph-loading')).not.toBeInTheDocument();
 
-    expect(mockEventStats).toHaveBeenCalledWith(
-      '/organizations/org-slug/events-stats/',
-      expect.objectContaining({
-        query: expect.objectContaining({
-          query: [persistantQuery, locationQuery.query.query].join(' '),
-        }),
+    expect(screen.getByRole('button', {name: 'All Envs'})).toBeInTheDocument();
+    expect(screen.getByRole('button', {name: '14D'})).toBeInTheDocument();
+    expect(screen.getByPlaceholderText('Filter events...')).toBeInTheDocument();
+    expect(
+      screen.getByRole('button', {
+        name: 'Toggle graph series - Events',
       })
-    );
-  });
-
-  it('allows filtering by date', async function () {
-    render(<EventDetailsHeader {...defaultProps} />, {organization});
-    await screen.findByRole('button', {
-      name: 'Toggle graph series - Events',
-    });
-
-    await userEvent.click(screen.getByRole('button', {name: '14D'}));
-    await userEvent.click(await screen.findByRole('option', {name: 'Last 7 days'}));
-
-    expect(mockEventStats).toHaveBeenCalledWith(
-      '/organizations/org-slug/events-stats/',
-      expect.objectContaining({
-        query: expect.objectContaining({
-          statsPeriod: '7d',
-        }),
-      })
-    );
-  });
-
-  it('displays error messages from bad queries', async function () {
-    const errorMessage = 'wrong, try again';
-    const mockStats = MockApiClient.addMockResponse({
-      url: `/organizations/${organization.slug}/events-stats/`,
-      body: {detail: errorMessage},
-      method: 'GET',
-      statusCode: 400,
-    });
-
-    render(<EventDetailsHeader {...defaultProps} />, {organization});
-    await screen.findByRole('button', {name: '14D'});
-
-    expect(mockStats).toHaveBeenCalled();
-    expect(screen.getByText(RegExp(errorMessage))).toBeInTheDocument();
-    // Omit the graph
-    expect(screen.queryByRole('figure')).not.toBeInTheDocument();
+    ).toBeInTheDocument();
+    expect(
+      screen.getByRole('button', {name: 'Toggle graph series - Users'})
+    ).toBeInTheDocument();
+    expect(screen.getByRole('figure')).toBeInTheDocument();
+    expect(screen.getByRole('button', {name: 'Open in Discover'})).toBeInTheDocument();
   });
 
   it('updates the query params with search tokens', async function () {
     const [tagKey, tagValue] = ['user.email', 'leander.rodrigues@sentry.io'];
     const locationQuery = {
       query: {
+        ...router.location.query,
         query: `${tagKey}:${tagValue}`,
       },
     };
@@ -216,12 +94,10 @@ describe('EventDetailsHeader', () => {
       method: 'GET',
     });
 
-    render(<EventDetailsHeader {...defaultProps} />, {organization});
-    await screen.findByRole('button', {
-      name: 'Toggle graph series - Events',
-    });
+    render(<EventDetailsHeader {...defaultProps} />, {organization, router});
+    expect(await screen.findByTestId('event-graph-loading')).not.toBeInTheDocument();
 
-    const search = screen.getAllByRole('combobox', {name: 'Add a search term'})[0];
+    const search = screen.getByPlaceholderText('Filter events...');
     await userEvent.type(search, `${tagKey}:`);
     await userEvent.keyboard(`${tagValue}{enter}{enter}`);
     expect(mockUseNavigate).toHaveBeenCalledWith(expect.objectContaining(locationQuery), {
